Moved the Engine config to lib
