+++ lib/buri_auth.rb
@@ -1,5 +1,9 @@
 # Override the standard devise views by requiring 'devise' before the engine
 require 'buri_auth/engine'
+require 'buri_auth/omniauth_helper'
+require 'buri_auth/devise_setup'
+require 'buri_auth/user'
+require 'buri_auth/identity'
 
 module BuriAuth
   class Configuration
@@ -24,5 +28,26 @@ module BuriAuth
 
   def setup
     yield(configuration)
+
+    after_setup
+  end
+
+  def after_setup
+    extend_models
+    setup_devise
+    # create_route_helpers
+  end
+
+  def setup_devise
+    BuriAuth::DeviseSetup.setup_with(configuration)
+  end
+
+  def extend_models
+    BuriAuth::User.extend_model
+    BuriAuth::Identity.extend_model
+  end
+
+  def create_route_helpers
+    configuration.omniauth_providers.each { |provider| BuriAuth::OmniauthHelper.configure_omniauth_provider(provider) }
   end
 end
